Handle non-GET requests with a generic response
Every incoming request gets written to the filesystem already. This is the first thing the server does when a request comes in. If this write fails, the server panics and does not complete the request, even if the request is read-only (aka GET method). The most likely failure state is runnning out of disk space. If this happens, all sites hosted on the server will become unresponsive. This is a good thing. We should be made aware right away so that we can quickly fix the problem. It should relatively easy to redeploy in read-only mode while the fix is underway. Sorry for the tangent.

The important thing to know is that we are already logging the entirety of the request. This allows us to design asynchronous systems to process these logs and apply any necessary changes to the system, which is the trend right now in the application space anyway.

Another added benifit of this architecture is that it doesn't leak implementation details. All the client needs to know is that the message got delivered and that they do not need to try again.